Build our own nip01 util
