+++ apps/arcade-city/package.json
@@ -41,6 +41,7 @@
     "@react-navigation/native-stack": "^6.1.0",
     "@react-spring/native": "9.5.2",
     "@sentry/react-native": "^4.1.3",
+    "create-hash": "1.2.0",
     "expo": "~46.0.0-beta.6",
     "expo-application": "~4.2.2",
     "expo-asset": "~8.6.1",
@@ -66,6 +67,7 @@
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
+    "@types/create-hash": "^1.2.2",
     "@types/react": "~18.0.14",
     "@types/react-native": "~0.69.1",
     "jest": "^26.6.3",
