+++ apps/arcade-city/src/lib/nostr/index.ts
@@ -1,1 +1,2 @@
 export * from './tools'
+export * from './types'
